+++ server/mailer/mailer_test.go
@@ -1,40 +1,14 @@
 package mailer_test
 
 import (
-	"context"
 	"fmt"
 	"testing"
 
 	"github.com/OrgaNiUS/OrgaNiUS/server/mailer"
-	"github.com/sendgrid/rest"
-	"github.com/sendgrid/sendgrid-go/helpers/mail"
 )
 
-type MockClient struct {
-	lastSend *mail.SGMailV3
-}
-
-func (c *MockClient) Send(email *mail.SGMailV3) (*rest.Response, error) {
-	c.lastSend = email
-	return nil, nil
-}
-
-func (c *MockClient) SendWithContext(ctx context.Context, email *mail.SGMailV3) (*rest.Response, error) {
-	// Not used.
-	return nil, nil
-}
-
-func getTestMailer() *mailer.Mailer {
-	name := "tester"
-	sender := "tester@test.com"
-	key := "abcdef"
-	mailer := mailer.New(name, sender, key)
-	mailer.Client = &MockClient{}
-	return mailer
-}
-
 func TestSend(t *testing.T) {
-	mailer := getTestMailer()
+	mail, mailer := mailer.GetTestMailer()
 
 	type sendData struct {
 		name, address, subject, body string
@@ -47,7 +21,6 @@ func TestSend(t *testing.T) {
 
 	for _, test := range tests {
 		mailer.Send(test.name, test.address, test.subject, test.body)
-		mail := mailer.Client.(*MockClient).lastSend
 		subject := mail.Subject
 		content := mail.Content
 		body := content[0].Value
@@ -61,7 +34,7 @@ func TestSend(t *testing.T) {
 }
 
 func TestSendVerification(t *testing.T) {
-	mailer_ := getTestMailer()
+	mail, mailer_ := mailer.GetTestMailer()
 
 	type sendData struct {
 		name, email, pin string
@@ -74,7 +47,6 @@ func TestSendVerification(t *testing.T) {
 
 	for _, test := range tests {
 		mailer_.SendVerification(test.name, test.email, test.pin)
-		mail := mailer_.Client.(*MockClient).lastSend
 		if mail.Subject != mailer.SignupSubject {
 			t.Errorf("Expected subject %v but got %v", mailer.SignupSubject, mail.Subject)
 		}
@@ -87,7 +59,7 @@ func TestSendVerification(t *testing.T) {
 }
 
 func TestSendForgotPW(t *testing.T) {
-	mailer_ := getTestMailer()
+	mail, mailer_ := mailer.GetTestMailer()
 
 	type sendData struct {
 		name, email, pin string
@@ -100,7 +72,6 @@ func TestSendForgotPW(t *testing.T) {
 
 	for _, test := range tests {
 		mailer_.SendForgotPW(test.name, test.email, test.pin)
-		mail := mailer_.Client.(*MockClient).lastSend
 		if mail.Subject != mailer.ForgotPWSubject {
 			t.Errorf("Expected subject %v but got %v", mailer.ForgotPWSubject, mail.Subject)
 		}
